Centralize all logging to `tracing` instead of `log` (#5670)
`tracing` has more structured logging support (see the changes I made in
`btrblocks`) which is more suitable for modern observability tooling. It
also has tracing support which is useful and already used in a few
places.

In python we use the tracing "log" feature which make it emit `log`
events

Signed-off-by: Adam Gutglick <adam@spiraldb.com>