Feature: Implement `take` on more vectors + stubs (#5653)
Tracking Issue: https://github.com/vortex-data/vortex/issues/5652

https://github.com/vortex-data/vortex/pull/5540 has to be merged first.

Implements `take` on the remaining vectors, except for `DecimalVector`,
`BinaryViewVector`, and `ListViewVector`

For `DecimalVector` we cannot trivially reuse the existing `take`
compute on slices and buffers because it has a bound of `NativePType`.
We can relax this bound to `Copy` instead, but that will require some
more work. Note that the current `DecimalArray` `take` implementation is
fully scalar right now so we can definitely improve that in the new
world.

For the view vectors I believe we want to think through how we "rebuild"
those in the array plan (we probably want to place rebuild nodes at good
places after just doing a `take` on the existing views).

So this PR implements:
- `FixedSizeListVector`
- `NullVector`
- `StructVector`

Edit: Now that https://github.com/vortex-data/vortex/pull/5666 has
merged, this also implements `DecimalVector`.

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>