Chore: Add all logical bool implementations + refactor (#5669)
This PR consolidates and finalizes a lot of the boolean and Kleene
logic, and also deduplicates a bunch of code.

Instead of having each of the operands in separate modules, I grouped
them into logical operands, kleene operands, and `not` by itself since
it is the only unary operator.

The convenience trait impls all have the same 3 impls: for
`&BoolScalar`, `&BoolVector`, `&BoolDatum`.

```rust
impl LogicalAndKleene for &BoolScalar { ... }

impl LogicalAndKleene for &BoolVector { ... }

impl LogicalAndKleene for &BoolDatum { ... }
```

Also adds more tests.

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>