Feature: implement `DictArray` batched execute (#5686)
Tracking Issue: https://github.com/vortex-data/vortex/issues/5652

Adds batched execution to `DictArray`, which is just implemented as
`take` on the child `values` vector with the codes as indices.

~~Additionally adds an optimization for `BoolVector::take` which is
similar to the optimization we have for `BoolArray`'s `canonicalize`
function, but does an additional check for zero or one `false`s (instead
of just zero or one `true`s). That code is located at
https://github.com/vortex-data/vortex/blob/develop/vortex-array/src/arrays/dict/vtable/canonical.rs.~~

~~The difference here is that I use a heuristic check on the default
`take` implementation on `BoolVector` (instead of only use this
optimization for dictionary decompression) because I don't think there
is any reason not to utilize this in general.~~

~~I still need to add some benchmarks.~~

Also note that we don't need to specialize for the string types like in
https://github.com/vortex-data/vortex/pull/1146 since we are _always_
decompressing the children immediately.

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>
Co-authored-by: Joe Isaacs <joe.isaacs@live.co.uk>