Feature: add some `Cast` infrastructure (#5708)
Adds the very basic implementations of `Cast` to `vortex-compute`.

The only things that are implemented are:
- Cast to `NullVector`/`NullScalar`
- Passthrough cast to the extension storage type
- Identity cast (a cast that is just cloning `self`)

All of the non-extremely-trivial cast implementations have not been
added, but this PR makes it a lot easier to see what is missing and a
lot easier to add those impls.

Note that I've also constrained the cast implementations so that we
cannot cast nullable data to a non-nullable `DType`, and this is just
checked by calling `self.validity().all_true()` on the vector which
should rarely be expensive.

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>