Perf: Optimized bool take (#5701)
Adds an optimization for `BoolVector::take` which is similar to the
optimization we have for `BoolArray`'s canonicalize function, but does
an additional check for zero or one `false`s (instead of just zero or
one `true`s). That code is located at
https://github.com/vortex-data/vortex/blob/develop/vortex-array/src/arrays/dict/vtable/canonical.rs.

The difference here is that I use a heuristic check on the default take
implementation on `BoolVector` (instead of only use this optimization
for dictionary decompression) because I don't think there is any reason
not to utilize this in general.

I still need to add some benchmarks.

---------

Signed-off-by: Connor Tsui <connor.tsui20@gmail.com>